Graph.jl

BayesNet.AbstractNode

Base.rand

BayesNet.add_logdensity

BayesNet.childnames

BayesNet.children

BayesNet.evaluate

BayesNet.merge_value

BayesNet.model

BayesNet.nodename

BayesNet.parents

BayesNet.prior

BayesNet.rng

BayesNet.traverse

Bijectors.bijector

DensityInterface.logdensityof

BayesNet.AbstractNode — Type

AbstractNode{name,child_names}

General graph traversing algorithms are defined on this type. name is typically a symbol and child_names a tuple of symbols.

If a node is a leaf (has nod children), it does not depend on any other variables and should have a fully specified model.

Base.rand — Method

rand(node, [variables, dims...])

Generate the random variables from the model by traversing the child nodes. Each node is evaluated only once and the dims are only applied to leafs. The variables parameter allows to condition the model and will not be re-sampled.

BayesNet.childnames — Method

childnames(node)

Extracts a tuple of the names of the child nodes.

BayesNet.children — Method

children(node)

Returns a tuple of the child nodes

BayesNet.evaluate — Method

evaluate(node, variables)

Evaluate only the deterministic nodes in the graph given the random variables. All required random variables are assumed to be available.

BayesNet.merge_value — Method

merge_value(variables, node, value)

Right to left merges the value for the node with the correct name into the previously sampled variables. Allows to override / modify previous values. If the value is nothing, the variable does not get merged

BayesNet.model — Method

model(node)

Returns the callable model of the node. Note that a Callable struct are not type stable and should be wrapped by a (anonymous) function.

BayesNet.nodename — Method

nodename(node)

Extracts the name of the node

BayesNet.parents — Method

parents(root::AbstractNode, node_name)

Returns a SequentializedGraph for the parents of the node_name node up until the root node.

BayesNet.prior — Method

prior(node)

The prior of a node are all the child nodes. Returns a SequentializedGraph for the prior

BayesNet.rng — Method

model(node)

Returns the random number generator of the node.

BayesNet.traverse — Method

traverse(fn, node, variables, [args...])

Effectively implements a depth first search to all nodes of the graph.

fn(node, variables, args...) is a function of the current node, the variables gathered from the recursions and args of the traverse function. The return values of fn are accumulated in a NamedTuple indexed by the node name. Only the first value of a node is considered, repeated calls for the same node name are ignored. If nothing is returned, the value is ignored. p…

Bijectors.bijector — Method

bijector(node)

Infer the bijectors of the model by traversing the child nodes. Internally a random is used to instantiate the models.

DensityInterface.logdensityof — Method

logdensityof(node, variables)

Calculate the logdensity of the model given the variables by traversing the child nodes. Each node is evaluated only once.

BayesNet.add_logdensity — Method

add_logdensity(A, B)

Add logdensities from different devices by converting them to CPU Arrays explicitly.