Predefined Node Types

BayesNet.BroadcastedNode

BayesNet.BroadcastedNode

BayesNet.BroadcastedNode

BayesNet.DeterministicNode

BayesNet.ModifierNode

BayesNet.SimpleNode

BayesNet.broadcast_model

BayesNet.insertdims

SimpleNode.jl

BayesNet.SimpleNode — Type

SimpleNode(name, rng, distribution_type, children)

Basic implementation of an AbstractNode with the main purpose of unit testing the graph library. Represents a named variable and depends on child nodes. Does not support logdensityof multiple samples, since no broadcasting or reduction is implemented.

BroadcastedNode.jl

BayesNet.BroadcastedNode — Type

BroadcastedNode

Broadcasts the parameters of the children using a BroadcastedDistribution. Also takes care of matching the dimensions for broadcasting multiple samples. model is a function which generates a BroadcastedDistribution given the nodes child values as parameters. child_sizes are the dims of a single sample from the child node and model_dims the dims of the broadcasted distribution which results from a single sample of all child nodes.

BayesNet.BroadcastedNode — Method

BroadcastedNode(name, rng, distribution, params...)

Construct the node as leaf (no children) by broadcasting the distribution over the params. The resulting BroadcastedDistribution acts like a product distribution, reducing the ndims of the params.

BayesNet.BroadcastedNode — Method

BroadcastedNode(name, rng, distribution, children)

Construct a node which automatically broadcasts the distribution over the parameters given by the children. The resulting BroadcastedDistribution acts like a product distribution, reducing the ndims for the minimal realization of the distribution given the children.

BayesNet.broadcast_model — Method

broadcast_model(fn, dims)

Wraps the distribution generating function with a BroadcastedDistribution given the reduction dims.

BayesNet.insertdims — Method

insertdims(A, child_size, model_dims)

Reshape A by inserting dims of length one to make it compatible for broadcasting multiple random samples of differently shaped children. child_size are the dims of a single sample from the child node and model_dims the dims of the broadcasted distribution which results from a single sample of all child nodes.

Rationale

When proposing multiple samples, originally matching dims of the BroadcastedDistribution do not work anymore. E.g. if one child has (3,) and the other (3,2) sized samples, proposing multiple samples result in incompatible dimensions (3,5) and (3,2,5). Julia expands dimensions of length one when broadcasting, so reshaping the array with dimensions of length one enables proposing multiple samples, for the above: (3,5) insertdims → (3,1,5) broadcast → (3,2,5) https://freecontent.manning.com/vectorizing-your-code-using-broadcasting/ https://discourse.julialang.org/t/designating-the-axes-for-broadcasting/29203/2

DeterministicNode.jl

BayesNet.DeterministicNode — Type

DeterministicNode

This node only takes part in the generative rand process but is not random. Instead a deterministic function fn is provided, e.g. rendering an image for a pose. It does not change the joint logdensity of the graph by returning a logdensity of zero.

ModifierNode.jl

BayesNet.ModifierNode — Type

ModifierNode

Wraps another node and represents the same variable as the wrapped node. For the model create a new type e.g. ModifierModel and implement:

rand(rng,::AbstractRNG, model::ModifierModel, x), where x is a randomly drawn value from the wrapped node.

logdensityof(model::ModifierModel, x, ℓ), where x is the value of the evaluated variable and ℓ the logdensity returned by the wrapped model

When traversing the graph, the children of the wrapped node are returned.